Organize test sources
