+++ test/src/commonTest/kotlin/io/github/kitakkun/kondition/test/requirement/FloatRequirementTest.kt
@@ -1,11 +1,5 @@
-package io.github.kitakkun.kondition.test
+package io.github.kitakkun.kondition.test.requirement
 
-import io.github.kitakkun.kondition.core.annotation.CoerceAtLeast
-import io.github.kitakkun.kondition.core.annotation.CoerceAtLeastDecimal
-import io.github.kitakkun.kondition.core.annotation.CoerceAtMost
-import io.github.kitakkun.kondition.core.annotation.CoerceAtMostDecimal
-import io.github.kitakkun.kondition.core.annotation.CoerceIn
-import io.github.kitakkun.kondition.core.annotation.CoerceInDecimal
 import io.github.kitakkun.kondition.core.annotation.GreaterThanDecimal
 import io.github.kitakkun.kondition.core.annotation.GreaterThanOrEqualsDecimal
 import io.github.kitakkun.kondition.core.annotation.LessThanDecimal
@@ -18,10 +12,9 @@ import io.github.kitakkun.kondition.core.annotation.Positive
 import io.github.kitakkun.kondition.core.annotation.RangeRule
 import io.github.kitakkun.kondition.core.annotation.RangedDecimal
 import kotlin.test.Test
-import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 
-class FloatTest {
+class FloatRequirementTest {
     @Test
     fun testRanged_InclusiveInclusive() {
         fun ranged(@RangedDecimal(start = 0.0, end = 10.0) value: Float) {
@@ -234,69 +227,5 @@ class FloatTest {
         assertFailsWith(IllegalArgumentException::class) {
             lessThanOrEquals(1.0f)
         }
-    }
-
-    @Test
-    fun testCoerceIn() {
-        fun coerceInUpper(@CoerceIn(0, 10) value: Float) {
-            assertEquals(10.0f, value)
-        }
-
-        fun coerceInLower(@CoerceIn(0, 10) value: Float) {
-            assertEquals(0.0f, value)
-        }
-
-        coerceInUpper(50.0f)
-        coerceInLower(-50.0f)
-    }
-
-    @Test
-    fun testCoerceAtMost() {
-        fun coerceAtMost(@CoerceAtMost(0) value: Float) {
-            assertEquals(0.0f, value)
-        }
-
-        coerceAtMost(50.0f)
-    }
-
-    @Test
-    fun testCoerceAtLeast() {
-        fun coerceAtLeast(@CoerceAtLeast(0) value: Float) {
-            assertEquals(0.0f, value)
-        }
-
-        coerceAtLeast(-50.0f)
-    }
-
-    @Test
-    fun testCoerceInDecimal() {
-        fun coerceInUpper(@CoerceInDecimal(0.0, 10.0) value: Float) {
-            assertEquals(10.0f, value)
-        }
-
-        fun coerceInLower(@CoerceInDecimal(0.0, 10.0) value: Float) {
-            assertEquals(0.0f, value)
-        }
-
-        coerceInUpper(10.1f)
-        coerceInLower(-0.1f)
-    }
-
-    @Test
-    fun testCoerceAtMostDecimal() {
-        fun coerceAtMost(@CoerceAtMostDecimal(0.0) value: Float) {
-            assertEquals(0.0f, value)
-        }
-
-        coerceAtMost(0.1f)
-    }
-
-    @Test
-    fun testCoerceAtLeastDecimal() {
-        fun coerceAtLeast(@CoerceAtLeastDecimal(0.0) value: Float) {
-            assertEquals(0.0f, value)
-        }
-
-        coerceAtLeast(-0.1f)
     }
 }
