+++ test/src/commonTest/kotlin/io/github/kitakkun/kondition/test/requirement/LongRequirementTest.kt
@@ -1,11 +1,5 @@
-package io.github.kitakkun.kondition.test
+package io.github.kitakkun.kondition.test.requirement
 
-import io.github.kitakkun.kondition.core.annotation.CoerceAtLeast
-import io.github.kitakkun.kondition.core.annotation.CoerceAtLeastDecimal
-import io.github.kitakkun.kondition.core.annotation.CoerceAtMost
-import io.github.kitakkun.kondition.core.annotation.CoerceAtMostDecimal
-import io.github.kitakkun.kondition.core.annotation.CoerceIn
-import io.github.kitakkun.kondition.core.annotation.CoerceInDecimal
 import io.github.kitakkun.kondition.core.annotation.GreaterThan
 import io.github.kitakkun.kondition.core.annotation.GreaterThanOrEquals
 import io.github.kitakkun.kondition.core.annotation.LessThan
@@ -18,10 +12,9 @@ import io.github.kitakkun.kondition.core.annotation.Positive
 import io.github.kitakkun.kondition.core.annotation.RangeRule
 import io.github.kitakkun.kondition.core.annotation.Ranged
 import kotlin.test.Test
-import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 
-class LongTest {
+class LongRequirementTest {
     @Test
     fun testRanged_InclusiveInclusive() {
         fun ranged(@Ranged(start = 0, end = 10) value: Long) {
@@ -234,69 +227,5 @@ class LongTest {
         assertFailsWith(IllegalArgumentException::class) {
             lessThanOrEquals(1)
         }
-    }
-
-    @Test
-    fun testCoerceIn() {
-        fun coerceInUpper(@CoerceIn(0, 10) value: Long) {
-            assertEquals(10, value)
-        }
-
-        fun coerceInLower(@CoerceIn(0, 10) value: Long) {
-            assertEquals(0, value)
-        }
-
-        coerceInUpper(50)
-        coerceInLower(-50)
-    }
-
-    @Test
-    fun testCoerceAtMost() {
-        fun coerceAtMost(@CoerceAtMost(0) value: Long) {
-            assertEquals(0, value)
-        }
-
-        coerceAtMost(50)
-    }
-
-    @Test
-    fun testCoerceAtLeast() {
-        fun coerceAtLeast(@CoerceAtLeast(0) value: Long) {
-            assertEquals(0, value)
-        }
-
-        coerceAtLeast(-50)
-    }
-
-    @Test
-    fun testCoerceInDecimal() {
-        fun coerceInUpper(@CoerceInDecimal(0.0, 10.0) value: Long) {
-            assertEquals(10, value)
-        }
-
-        fun coerceInLower(@CoerceInDecimal(0.0, 10.0) value: Long) {
-            assertEquals(0, value)
-        }
-
-        coerceInUpper(50)
-        coerceInLower(-50)
-    }
-
-    @Test
-    fun testCoerceAtMostDecimal() {
-        fun coerceAtMost(@CoerceAtMostDecimal(0.0) value: Long) {
-            assertEquals(0, value)
-        }
-
-        coerceAtMost(50)
-    }
-
-    @Test
-    fun testCoerceAtLeastDecimal() {
-        fun coerceAtLeast(@CoerceAtLeastDecimal(0.0) value: Long) {
-            assertEquals(0, value)
-        }
-
-        coerceAtLeast(-50)
     }
 }
